+++ src/libtiledb.cpp
@@ -26,6 +26,7 @@
 #include <fstream>
 #include <algorithm>
 #include <unistd.h>
+#include <optional>
 
 using namespace Rcpp;
 
@@ -5925,3 +5926,73 @@ bool libtiledb_current_domain_is_empty(XPtr<tiledb::CurrentDomain> cd) {
   return false;
 #endif
 }
+
+
+/**
+ * Profile (2.29.0 or later)
+ */
+
+std::optional<std::string> as_optional_string(Nullable<CharacterVector> param) {
+    if (param.isNull()) {
+        return std::nullopt;
+    }
+    CharacterVector tmp(param);
+    return std::string(tmp[0]);
+}
+
+
+// [[Rcpp::export]]
+XPtr<tiledb::Profile> libtiledb_profile_new(Nullable<CharacterVector> name = R_NilValue, Nullable<CharacterVector> dir  = R_NilValue) {
+    auto name_ = as_optional_string(name);
+    auto dir_ = as_optional_string(dir);
+    return make_xptr<tiledb::Profile>(new tiledb::Profile(name_, dir_));
+}
+
+// [[Rcpp::export]]
+XPtr<tiledb::Profile> libtiledb_profile_load(Nullable<CharacterVector> name = R_NilValue, Nullable<CharacterVector> dir = R_NilValue) {
+    auto name_ = as_optional_string(name);
+    auto dir_ = as_optional_string(dir);
+    return make_xptr<tiledb::Profile>(new tiledb::Profile(tiledb::Profile::load(name_, dir_)));
+}
+
+// [[Rcpp::export]]
+void libtiledb_profile_remove(Nullable<CharacterVector> name = R_NilValue, Nullable<CharacterVector> dir = R_NilValue) {
+    auto name_ = as_optional_string(name);
+    auto dir_ = as_optional_string(dir);
+    return tiledb::Profile::remove(name_, dir_);
+}
+
+// [[Rcpp::export]]
+std::string libtiledb_profile_name(XPtr<tiledb::Profile> profile) {
+    return profile->name();
+}
+
+// [[Rcpp::export]]
+std::string libtiledb_profile_dir(XPtr<tiledb::Profile> profile) {
+    return profile->dir();
+}
+
+// [[Rcpp::export]]
+void libtiledb_profile_set_param(XPtr<tiledb::Profile> profile, const std::string& param, const std::string& value) {
+    return profile->set_param(param, value);
+}
+
+// [[Rcpp::export]]
+Nullable<CharacterVector> libtiledb_profile_get_param(XPtr<tiledb::Profile> profile, const std::string& param) {
+    auto maybe_value = profile->get_param(param);
+    if (maybe_value.has_value()) {
+        return wrap<CharacterVector>(maybe_value.value());
+    }
+    return R_NilValue;
+}
+
+// [[Rcpp::export]]
+void libtiledb_profile_save(XPtr<tiledb::Profile> profile) {
+    return profile->save();
+}
+
+// [[Rcpp::export]]
+std::string libtiledb_profile_dump(XPtr<tiledb::Profile> profile) {
+    return profile->dump();
+}
+