+++ NEWS.md
@@ -1,5 +1,9 @@
 # tiledb (development version)
 
+## Improvements
+
+* Add support for creating, loading, saving, and removing Profiles ([#856](https://github.com/TileDB-Inc/TileDB-R/pull/856))
+
 # tiledb 0.33.1
 
 * This release of the R package builds against [TileDB 2.29.1](https://github.com/TileDB-Inc/TileDB/releases/tag/2.29.1), and has also been tested against earlier releases as well as the development version